Create Problem_1
This was the first problem involving Fibonacci numbers. The program prompts a user to enter a number less than or equal to nine and then displays that amount of numbers. If the user decides to enter a larger number than nine, it disregards and only prints the nine numbers. The output is as follows:
How many numbers would you like to print? (less than 9)
15
1 1 2 3 5 8 13 21 34 
-- program is finished running --