Updated_Problem 2
This program takes two integers from the user and prints out the addition, subtraction, and production of the two numbers. It prints out a brief introduction, the results of the calculations, and a brief outro. The result look like :

Name: Justin Hiller
Simple Calculator:
This program takes two integers from the user and performs calculations on the numbers.:
Enter a number: 16
Enter a number: 5

The sum of the numbers is: 21
The difference of the numbers is: 11
The product of the numbers is: 80
Thanks for playing! Goodbye
-- program is finished running --