Justin Hiller Problem 1
This program calculates the first nine Fibonacci numbers and displays them. The program takes in an input from the user asking how many numbers are to be displayed. The user cannot enter a number greater than nine. If the user decides to enter a number greater than nine, it only prints out the nine numbers as follows.

How many numbers would you like to print? (less than 9)
15
1 1 2 3 5 8 13 21 34 
-- program is finished running --

This is the output if the user submits a number within the range:

How many numbers would you like to print? (less than 9)
4
The Fibonacci numbers are:
1 1 2 3 
-- program is finished running --