Justin Hiller Problem 2
This program takes two numbers from the user and then performs calculations on those numbers. It performs an addition, a subtraction, and a production of the two numbers. The output is as follows:

Name: Justin Hiller
Simple Calculator:
This program takes two integers from the user and performs calculations on the numbers.:
Enter a number: 15
Enter a number: 6

The sum of the numbers is: 21
The difference of the numbers is: 9
The product of the numbers is: 90
Thanks for playing! Goodbye
-- program is finished running --